GO code Array
Generates an array GOArray which has 3 columns and N rows, where N is
equal to the number of proteins in OverlordMatrix (lacking decoy
proteins).  Column 1 are the string GO codes, column 2 are the GO code
types in order (biological, molecular, cellular), and column 3 are the
numerical forms of the GO codes in column 1.  The indeces of the GOArray
correspond to those in 'axes' from OverlordMatrix, i.e. row 2 of GOArray
corresponds to protein 2.  This means that there are some empty rows in
GOArray for proteins lacking GO codes, e.g. protein 1.