Final dataset.  Done.  Finished.  veni vidi vici.
The new excel file has all '##' terms removed and has 'Final' to the
right of 'NoDecoy' in the name.  Changed the initialization script to
call this new excel sheet.

Modified the script for OverlordNormalizer just a touch--no longer asks
for a choice of 1 or 2 and instead simply runs each normalization
sequentially.  The final output matrix everyone should use always is
named 'normOverlordF', F for final, not for horrible horrible failure.
Hopefully.  It's 6:17 AM and I'm still up.  I've lost control of my
life.  gg.